Rework category init and add custom (Fixed #5)
Now we have more categories: custom and mods. Mods isn't used for now. Custom is for people who wants to add custom musics to minecraft via datapacks.
The ids are now split with namespace and id so that it's more clear. There is also a mod list to know how many musics a mod have (will be used later too).
Changed how musics are played so that we can play musics that aren't registered (such as custom ones).
When reloading minecraft (resourcepack for example), we re init the categories.