Print the right name of musics (Fixed #8)
Added the name of every music in the lang file to see their real name instead of a path. For custom songs, this won't work as I don't have their names (unless they add the right thing in their lang file). Instead, the path will be displayed as before. Also, the display is not adaptive for cat/title/pause