Use biome/soundEvent map instead of optional sound
We can't register biome modifications when playing in server and we want
to keep the mod client only, so we use our own music/biome map. It
doesn't have to know modded biomes for their music, as it will use the
biome music if not found in the custom map, which is only to get events
for biomes for which we registered new sound events.
Fixes #43