Fix potential crash and wrong music
When in an old growth taiga without resource pack enabled, it plays the default event. This prevents the music to be put in the history, so we use the biome map instead, as if rp was enabled.
In case we would still encounter issues when using the previous key in such biome, use a more robust check and plays the same sound event in the worst case.